+++ lib/src/annotations/annotations.dart
@@ -1,6 +1,7 @@
 library jaguar_serializer.annotations;
 
-import 'package:jaguar_serializer/src/serializer/serializer.dart';
+import '../serializer/serializer.dart';
+import '../serializer/custom_codec.dart';
 
 /// Annotation used to request generation of serializer
 class GenSerializer {
@@ -19,47 +20,109 @@ class GenSerializer {
   final List<String> ignore;
 
   /// [FieldProcessor] that shall be used to encode and decode specified property
+  @Deprecated("use fields property")
   final Map<String, FieldProcessor> processors;
 
   /// Supplies [Serializer]s to use when unknown PODO type is encountered
   final List<Type> serializers;
 
+  /// Enable null check on fields
+  final bool nullableFields;
+
   const GenSerializer(
-      {this.fields: const {},
-      this.ignore: const [],
-      this.processors: const {},
-      this.serializers: const [],
+      {this.fields: const <String, Property>{},
+      this.ignore: const <String>[],
+      this.processors: const <String, FieldProcessor>{},
+      this.serializers: const <Type>[],
       this.modelName,
-      this.includeByDefault: true});
+      this.includeByDefault: true,
+      this.nullableFields: false});
 }
 
-abstract class Property {}
+class Property<T> {
+  final String encodeTo;
+  final String decodeFrom;
+  final bool isNullable;
+  final FieldProcessor<T, dynamic> processor;
+  final T defaultsTo;
+  final bool valueFromConstructor;
 
+  const Property(
+      {this.encodeTo,
+      this.decodeFrom,
+      this.isNullable,
+      this.defaultsTo,
+      this.processor,
+      this.valueFromConstructor});
+}
+
+// can't use inheritance here, [DartObject.getField] does not support getter, only fields
 /// Annotation used to request encoding of a field in model
-class EncodeOnly implements Property {
+class EncodeOnly<T> extends Property<T> {
   /// Optional. Key used to encode the model in the [Map]
   final String alias;
+  final FieldProcessor<T, dynamic> processor;
+  final bool isNullable;
+  final T defaultsTo;
+  final String encodeTo;
+  final bool valueFromConstructor;
 
-  const EncodeOnly([this.alias]);
+  const EncodeOnly(
+      {this.alias,
+      this.isNullable,
+      this.defaultsTo,
+      this.processor,
+      this.valueFromConstructor})
+      : encodeTo = alias;
 }
 
 /// Annotation used to request decoding of a field in model
-class DecodeOnly implements Property {
+class DecodeOnly<T> extends Property<T> {
   /// Optional. Key used to decode the model from the [Map]
   final String alias;
+  final FieldProcessor<T, dynamic> processor;
+  final bool isNullable;
+  final T defaultsTo;
+  final String decodeFrom;
+  final bool valueFromConstructor;
 
-  const DecodeOnly([this.alias]);
+  const DecodeOnly(
+      {this.alias,
+      this.isNullable,
+      this.defaultsTo,
+      this.processor,
+      this.valueFromConstructor})
+      : decodeFrom = alias;
 }
 
 /// Annotation used to request encoding and decoding of a field in model
-class EnDecode implements Property {
+class EnDecode<T> extends Property<T> {
   /// Optional. Key used to decode and encode the model from and to the [Map]
   final String alias;
+  final FieldProcessor<T, dynamic> processor;
+  final bool isNullable;
+  final T defaultsTo;
+  final String encodeTo;
+  final String decodeFrom;
+  final bool valueFromConstructor;
 
-  const EnDecode([this.alias]);
+  const EnDecode(
+      {this.alias,
+      this.isNullable,
+      this.defaultsTo,
+      this.processor,
+      this.valueFromConstructor})
+      : encodeTo = alias,
+        decodeFrom = alias;
 }
 
 /// Annotation to ignore a field while encoding or decoding
-class Ignore implements Property {
+class Ignore extends EnDecode {
   const Ignore();
 }
+
+const ignore = const Ignore();
+const Property nullable = const Property(isNullable: true);
+const Property nonNullable = const Property(isNullable: false);
+const Property useConstructorForDefaultsValue =
+    const Property(valueFromConstructor: true, isNullable: false);