+++ lib/src/json/json.dart
@@ -1,14 +1,19 @@
 library jaguar_serializer.json;
 
 import 'dart:convert';
-import 'package:jaguar_serializer/jaguar_serializer.dart';
+
+import '../serializer/serializer.dart';
+import '../serializer/repo.dart';
 
 /// Repository that serialize/deserialize JSON.
 ///
 /// Same usage as [SerializerRepo]
 class JsonRepo extends SerializerRepo {
-  JsonRepo({List<Serializer> serializers, String typeKey: defaultTypeInfoKey})
-      : super(serializers: serializers, typeKey: typeKey);
+  JsonRepo(
+      {List<Serializer> serializers,
+      String typeKey: defaultTypeInfoKey,
+      bool withType})
+      : super(serializers: serializers, typeKey: typeKey, withType: withType);
 
   ///@nodoc
   dynamic encode(dynamic object) => JSON.encode(object);
@@ -31,6 +36,6 @@ class JsonRepo extends SerializerRepo {
   ///
   /// See [SerializerRepo.to] for more information.
   @override
-  dynamic serialize(dynamic object, {bool withType: false, String typeKey}) =>
+  dynamic serialize(dynamic object, {bool withType, String typeKey}) =>
       super.serialize(object, withType: withType, typeKey: typeKey);
 }