+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.5.0
+
+- support disable null check
+- deprecate `GenSerializer.processor` fields, FieldProcessor go into `GenSerializer.fields` inside `Property`
+- support for default value, `String`, `int`, `double`, `bool`, if a field is non nullable, jaguar will use the default value declared in you object constructor
+
+***Breaking Change***
+`EnDecode`, `DecodeOnly`, `EncodeOnly` does not use optional parameter in constructor anymore,
+
+example: `EnDecode({this.alias, bool isNullable, FieldProcessor<T, dynamic> processor, T defaultsTo})`
+
+## 0.4.2
+- Better support of ddc
+
 ## 0.4.1
 
 - Added docs
@@ -7,7 +21,24 @@
 
 ## 0.4.0
 
-- Angular style annotations
+***Breaking Changes***
+- Generator is now in a separate [package](https://pub.dartlang.org/packages/jaguar_serializer_cli)
+- Angular style annotations, only one annotation
+```dart
+@GenSerializer(
+  fields: const {
+    'field': const EncodeOnly('f')
+  }, 
+  ignore: const [
+    'ignoreField'
+  ], 
+  processors: const {
+   'dates': const DateTimeProcessor(),
+  }, 
+  serializers: const [
+   FieldSerializer,
+])
+```
 
 ## 0.3.16